% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_utils.R
\name{alphabetize_cols}
\alias{alphabetize_cols}
\title{Alphabetically sorts entries in data frame across two columns}
\usage{
alphabetize_cols(df, x, y)
}
\arguments{
\item{df}{data frame}

\item{x}{Column x}

\item{y}{Column y}
}
\value{
Data frame with same structure as input
}
\description{
Ensures that elements in column x are alphabetically before
elements in column y. If they are not, the entries are swapped.
}
