% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labdata.R
\name{bca_quant}
\alias{bca_quant}
\title{run sample quantification on plate reader data from BCA kit}
\usage{
bca_quant(f, dilution_factor = 5, target_amt = 20, target_vol = 15)
}
\arguments{
\item{f}{a csv file with header row including Sample, Value, and Conc columns}

\item{dilution_factor}{dilution factor of test samples}

\item{target_amt}{ug of sample to load in a gel}

\item{target_vol}{ul to load}
}
\value{
list with a dataframe and a ggplot object
}
\description{
Specify standards with "Standard" in Sample columns
and the concentrations in the Conc column
}
