% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio.R
\name{de_excelify_genes}
\alias{de_excelify_genes}
\title{Attempts to reverse adverse effects of opening a table in Excel.}
\usage{
de_excelify_genes(df, x = Gene)
}
\arguments{
\item{df}{data frame}

\item{x}{gene column name}
}
\value{
data frame with fixed gene column
}
\description{
Reverts gene names that have been converted to dates by Excel.
}
