% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{format_si}
\alias{format_si}
\title{Format a vector of numeric values according
to the International System of Units.
http://en.wikipedia.org/wiki/SI_prefix}
\usage{
format_si(...)
}
\arguments{
\item{...}{args passed to format()}
}
\value{
A function to format a vector of strings using SI prefix notation
}
\description{
Reverts gene names that have been converted to dates by Excel.
}
\details{
Based on code by Ben Tupper
https://stat.ethz.ch/pipermail/r-help/2012-January/299804.html
}
