% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{pretty_log10_limits}
\alias{pretty_log10_limits}
\title{Pretty limits for log10 scale}
\usage{
pretty_log10_limits(x)
}
\arguments{
\item{x}{numeric vector to find pretty log10 limits outside of range}
}
\value{
length two vector of limits
}
\description{
Pretty limits for log10 scale
}
