% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{read_gct}
\alias{read_gct}
\title{read a *.gct[.gz] file into your session}
\usage{
read_gct(gct_file, chunk_size = 1000)
}
\arguments{
\item{gct_file}{path to gct formatted file}

\item{chunk_size}{number of lines to read at once}
}
\value{
matrix with a "Description" attribute
}
\description{
read a *.gct[.gz] file into your session
}
