% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_utils.R
\name{row_correlations}
\alias{row_correlations}
\title{Correlation matrix}
\usage{
row_correlations(x, ...)
}
\arguments{
\item{x}{data matrix}

\item{...}{arguments passed to plyr::aaply()}
}
\value{
correlation matrix
}
\description{
Calculates correlations between all rows of a matrix.
Parallelizable.
}
